feat(idp): add app login service

+++ idp/internal/services/dtos/auth.go
@@ -9,7 +9,7 @@ package dtos
 type AuthDTO struct {
 	AccessToken  string            `json:"access_token"`
 	RefreshToken string            `json:"refresh_token,omitempty"`
-	ExpiresIn    int               `json:"expires_in"`
+	ExpiresIn    int64             `json:"expires_in"`
 	TokenType    string            `json:"token_type"`
 	Message      string            `json:"message,omitempty"`
 	Data         map[string]string `json:"data,omitempty"`
@@ -20,7 +20,7 @@ const tokenType string = "Bearer"
 func NewAuthDTO(accessToken string, expiresIn int64) AuthDTO {
 	return AuthDTO{
 		AccessToken: accessToken,
-		ExpiresIn:   int(expiresIn),
+		ExpiresIn:   expiresIn,
 		TokenType:   tokenType,
 	}
 }
@@ -29,7 +29,7 @@ func NewFullAuthDTO(accessToken, refreshToken string, expiresIn int64) AuthDTO {
 	return AuthDTO{
 		AccessToken:  accessToken,
 		RefreshToken: refreshToken,
-		ExpiresIn:    int(expiresIn),
+		ExpiresIn:    expiresIn,
 		TokenType:    tokenType,
 	}
 }
@@ -37,7 +37,7 @@ func NewFullAuthDTO(accessToken, refreshToken string, expiresIn int64) AuthDTO {
 func NewTempAuthDTO(accessToken, message string, expiresIn int64) AuthDTO {
 	return AuthDTO{
 		AccessToken: accessToken,
-		ExpiresIn:   int(expiresIn),
+		ExpiresIn:   expiresIn,
 		TokenType:   tokenType,
 		Message:     message,
 	}
@@ -46,7 +46,7 @@ func NewTempAuthDTO(accessToken, message string, expiresIn int64) AuthDTO {
 func NewAuthDTOWithData(accessToken, message string, data map[string]string, expiresIn int64) AuthDTO {
 	return AuthDTO{
 		AccessToken: accessToken,
-		ExpiresIn:   int(expiresIn),
+		ExpiresIn:   expiresIn,
 		TokenType:   tokenType,
 		Message:     message,
 		Data:        data,
